CodeRabbit指摘に対応しapp_launch送信失敗時の再送を可能にしテストセットアップを共通化
